+++ app/Http/Controllers/PlaceController.php
@@ -70,6 +70,7 @@ class PlaceController extends Controller
             'lat' => ! isset($request->addressjson['latlng']['lat']) ? null : $request->addressjson['latlng']['lat'],
             'instructions' => $request->instructions,
             'deliveryZone' => $request->deliveryZone,
+            'hide_address' => $request->boolean('hideAddress'),
         ]);
 
         $place->categories()->sync($request->categories);
@@ -98,6 +99,7 @@ class PlaceController extends Controller
             'lat' => ! isset($request->addressjson['latlng']['lat']) ? null : $request->addressjson['latlng']['lat'],
             'instructions' => $request->instructions,
             'deliveryZone' => $request->deliveryZone,
+            'hide_address' => $request->boolean('hideAddress'),
         ]);
 
         $place->categories()->sync($request->categories);
@@ -160,6 +162,7 @@ class PlaceController extends Controller
         $place->additionnalPhoneNumber = $request->additionnalPhoneNumber;
         $place->email = $request->email;
         $place->url = $request->url;
+        $place->hide_address = $request->boolean('hideAddress');
         $place->deliveryZone = $request->deliveryZone;
         $place->save();