+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.14.0 (2020-03-28)
+- Display region on place.show.
+- Added ability to hide address from places.
+
 ## 1.13.0 (2020-03-28)
 - Added docker file for devs
 - Precised that "name" is the name of the place in the "add place" page, because people 🤷‍♂️